refactor(sync): split engine.test.ts into 4 domain files + shared tests/utils/engine-mocks.ts
Per .claude/plans/tech-debt-remediation.md Phase 3.2. The 3,136-line
engine.test.ts (74 describe blocks, 85 tests) is split by domain. Shared
scaffolding moves into tests/utils/engine-mocks.ts (follows existing
convention alongside test-db.ts, test-vault.ts, etc.). engine.ts source
is untouched.

- tests/utils/engine-mocks.ts (73 LOC): createMockNetwork, createMockWs,
  createMockDeps, setupTestDb — shared by every split file
- engine-pull.test.ts (528 LOC): pull + decrypt + unknown signer (9 tests)
- engine-push.test.ts (983 LOC): push + clock skew + buildPushPayload +
  duplicate queue + replay detection (22 tests; disambiguated duplicate titles)
- engine-crdt.test.ts (330 LOC): crdtProvider CREATE/UPDATE note + task,
  snapshot-fails, mixed batch (6 tests)
- engine-retries.test.ts (542 LOC): online/offline transitions, device
  revoked, extended offline container (21 tests)
- engine.test.ts (residual 599 LOC): lifecycle, fullSync, manifest,
  remote-wipe T245k (27 tests)

Total 85 tests preserved. Runtime behavior unchanged — this is pure
test-file reorganization. Location in tests/utils/ (rather than next to
engine.ts) keeps the helper in the tests/ tsconfig project so external
imports (e.g. @tests/utils/test-db for TestDatabaseResult) resolve
without needing tsconfig exclude workarounds.
